ReceiveStrategy implementations moved to separate files. Handling of the vibility timeout expiration now leaves a specific log entry with WARN lvl.
