+++ src/Transport/NServiceBus.AzureStorageQueues.csproj
@@ -101,6 +101,8 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AtLeastOnceReceiveStrategy.cs" />
+    <Compile Include="AtMostOnceReceiveStrategy.cs" />
     <Compile Include="AzureMessageQueueCreator.cs" />
     <Compile Include="AzureStorageQueueInfrastructure.cs" />
     <Compile Include="Config\Extensions\AccountConfigurations.cs" />
